refactor: 功能用例导入前端重构
