+++ frontend/src/business/components/track/case/components/import/TestCaseImport.vue
@@ -1,0 +1,82 @@
+<template>
+  <el-dialog class="testcase-import" :title="$t('test_track.case.import.case_import')" :visible.sync="dialogVisible" @close="close">
+    <el-tabs v-model="activeName" simple>
+
+      <el-tab-pane :label="$t('test_track.case.import.excel_title')" name="excelImport">
+        <test-case-common-import
+          :is-updated="isUpdated"
+          name="excel"
+          tab-name="excelImport"
+          @fresh="$emit('refreshAll')"
+          @close="close"
+          ref="excelImport"/>
+      </el-tab-pane>
+
+      <el-tab-pane :label="$t('test_track.case.import.xmind_title')" name="xmindImport">
+        <test-case-common-import
+          :is-updated="isUpdated"
+          name="xmind"
+          tab-name="xmindImport"
+          @fresh="$emit('refreshAll')"
+          @close="close"
+          ref="xmindImport"/>
+      </el-tab-pane>
+
+    </el-tabs>
+  </el-dialog>
+</template>
+
+<script>
+  import MsTableButton from '../../../../common/components/MsTableButton';
+  import {getCurrentProjectID, listenGoBack, removeGoBackListener} from "../../../../../../common/js/utils";
+  import TestCaseCommonImport from "@/business/components/track/case/components/import/TestCaseCommonImport";
+
+  export default {
+    name: "TestCaseImport",
+    components: {TestCaseCommonImport, MsTableButton},
+    data() {
+      return {
+        activeName: 'excelImport',
+        dialogVisible: false,
+        isLoading: false,
+        isUpdated: false,
+      }
+    },
+    methods: {
+      open() {
+        listenGoBack(this.close);
+        this.projectId = getCurrentProjectID();
+        this.dialogVisible = true;
+        this.init();
+      },
+      init() {
+        if (this.$refs.excelImport) {
+          this.$refs.excelImport.init();
+        }
+        if (this.$refs.xmindImport) {
+          this.$refs.xmindImport.init();
+        }
+      },
+      close() {
+        removeGoBackListener(this.close);
+        this.dialogVisible = false;
+        //通过excel导入更新过数据的话就刷新页面
+        if (this.isUpdated === true) {
+          this.$emit("refreshAll");
+          this.isUpdated = false;
+        }
+      },
+    }
+  }
+</script>
+
+<style>
+</style>
+
+<style scoped>
+
+.testcase-import >>> .el-dialog {
+  width: 650px;
+}
+
+</style>
